+++ tests/test_core.py
@@ -48,7 +48,7 @@ def test_bigchain_class_default_initialization(config):
 
 
 @pytest.mark.bdb
-def test_get_spent_issue_1271(b, alice, bob, carol):
+def test_get_spending_transaction_issue_1271(b, alice, bob, carol):
     tx_1 = Create.generate(
         [carol.public_key],
         [([carol.public_key], 8)],
@@ -88,7 +88,7 @@ def test_get_spent_issue_1271(b, alice, bob, carol):
     assert b.validate_transaction(tx_5)
 
     b.models.store_bulk_transactions([tx_5])
-    assert b.models.get_spent(tx_2.id, 0) == tx_5.to_dict()
-    assert not b.models.get_spent(tx_5.id, 0)
+    assert b.models.get_spending_transaction(tx_2.id, 0) == tx_5.to_dict()
+    assert not b.models.get_spending_transaction(tx_5.id, 0)
     assert b.models.get_outputs_filtered(alice.public_key)
     assert b.models.get_outputs_filtered(alice.public_key, spent=False)